+++ Disciple/Changes.md
@@ -3,6 +3,8 @@
 ## v1.9.3 (in development)
 ### Breaking changes
 - Updated to **Apache HttpClient v5.6**
+### New features
+- **RequestRateLimiter** now supports temporary lockdowns using `.lockUntil(Instant)` and `.removeLocking()`
 
 ## v1.9.2 - 15.03.2026
 A minor update adding some quality-of-life changes, and support for some **Echo**-specific use-cases.